backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers>=4.40.0
torch>=2.5.0
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentencepiece==0.1.99
accelerate>=1.0.0
google-generativeai>=0.3.0
protobuf>=3.20.0
